+++ app/Http/Middleware/VerifyCsrfToken.php
@@ -38,6 +38,6 @@ class VerifyCsrfToken extends Middleware
      * @var array
      */
     protected $except = [
-        //
+        'autoimport',
     ];
 }