+++ changelog.md
@@ -2,6 +2,14 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## 2.2.1 - 2020-10-04
+
+### Added
+- New `/autoimport` route you can POST to.
+
+### Fixed
+- A bug where JSON errors would lead to errors in the CSV importer.
+
 ## 2.2.0 - 2020-09-21
 
 ⚠️ Several changes in this release may break Firefly III's duplication detection or are backwards incompatible.